Move mailout functionality to nectar-osc
The first version will be a transplant of the hivemind_contrib
freshdesk notification command, but using Taynac rather than
calling the Freshdesk API directly.  The old (incorrect) use of
'--no-dry-run' flag to distinguish the 'prepare' and 'send' phase
has been replaced with separate subcommands; e.g.

  nectar mailout instances ...                 # prepare
  nectar mailout send --mailout-dir <dir> ...  # send
  nectar mailout clean --mailout-dir <dir>

The notifications are prepared with sequence numbers, and 'send'
keeps track of the sequence number last (successfully) sent ...
so that it will easy to restart after a failure during sending.

- SMTP email and direct Freshdeck are not supported
- Templates must be HTML.  (No automatic text -> HTML 'conversion'.  The
  old hivemind conversion code is too crude.)
- Notifications for volumes and VDS desktops are not supported yet.
- There are issues with '--limit' to resolve
- Need some feedback on the subcommands; e.g.
  - Should we use generated tmp dir names, or pass a name as an option?
  - Do we want to be able to add Freshdesk tags (via Taynac)?
- Need to convert hivemind_contrib's templates to HTML and add them
- Want to implement template search path

Change-Id: If08dee0f151d2807d75d2b4a597069a4efe46ac6